Created login controller endpoint and removed jwt package from middleware
