+++ shared/adapter/adapter.go
@@ -2,12 +2,12 @@ package adapter
 
 import (
 	"fmt"
+	"github.com/apostrohedottilde/indymorning/api/shared"
 	"net/http"
 
 	"github.com/apostrohedottilde/indymorning/api/project"
 	"github.com/apostrohedottilde/indymorning/api/user"
 
-	"github.com/apostrohedottilde/indymorning/api/shared/middleware/jwt"
 	l "github.com/apostrohedottilde/indymorning/api/shared/middleware/logger"
 	t "github.com/apostrohedottilde/indymorning/api/shared/middleware/terminator"
 
@@ -39,20 +39,20 @@ func (adapter *HTTPAdapter) Stop() {
 func New(u *user.UserController, p *project.ProjectController) *HTTPAdapter {
 	r := mux.NewRouter()
 
-	r.HandleFunc("/auth/login", l.Log(jwt.Sign(t.End())).ServeHTTP).Methods("POST")
+	r.HandleFunc("/auth/login", l.Log(u.Login(t.End())).ServeHTTP).Methods("POST")
 	r.HandleFunc("/auth/register", u.Register(t.End()).ServeHTTP).Methods("POST")
 
-	r.HandleFunc("/users/{id}", l.Log(jwt.Validate(u.FindOne(t.End()))).ServeHTTP).Methods("GET")
-	r.HandleFunc("/users/{id}", l.Log(jwt.Validate(u.Update(t.End()))).ServeHTTP).Methods("PUT")
-	r.HandleFunc("/users", l.Log(jwt.Validate(u.FindAll(t.End()))).ServeHTTP).Methods("GET")
-	r.HandleFunc("/users/{id}", l.Log(jwt.Validate(u.Delete(t.End()))).ServeHTTP).Methods("DELETE")
+	r.HandleFunc("/users/{id}", l.Log(shared.Validate(u.FindOne(t.End()))).ServeHTTP).Methods("GET")
+	r.HandleFunc("/users/{id}", l.Log(shared.Validate(u.Update(t.End()))).ServeHTTP).Methods("PUT")
+	r.HandleFunc("/users", l.Log(shared.Validate(u.FindAll(t.End()))).ServeHTTP).Methods("GET")
+	r.HandleFunc("/users/{id}", l.Log(shared.Validate(u.Delete(t.End()))).ServeHTTP).Methods("DELETE")
 
-	r.HandleFunc("/projects/{id}", l.Log(jwt.Validate(p.FindOne(t.End()))).ServeHTTP).Methods("GET")
-	r.HandleFunc("/projects/{id}", l.Log(jwt.Validate(p.Update(t.End()))).ServeHTTP).Methods("PUT")
-	r.HandleFunc("/projects", l.Log(jwt.Validate(p.FindAll(t.End()))).ServeHTTP).Methods("GET")
-	r.HandleFunc("/projects", l.Log(jwt.Validate(p.Create(t.End()))).ServeHTTP).Methods("POST")
-	r.HandleFunc("/projects/{id}", l.Log(jwt.Validate(p.Delete(t.End()))).ServeHTTP).Methods("DELETE")
-	r.HandleFunc("/projects/{id}/cancel", l.Log(jwt.Validate(p.Cancel(t.End()))).ServeHTTP).Methods("POST")
+	r.HandleFunc("/projects/{id}", l.Log(shared.Validate(p.FindOne(t.End()))).ServeHTTP).Methods("GET")
+	r.HandleFunc("/projects/{id}", l.Log(shared.Validate(p.Update(t.End()))).ServeHTTP).Methods("PUT")
+	r.HandleFunc("/projects", l.Log(shared.Validate(p.FindAll(t.End()))).ServeHTTP).Methods("GET")
+	r.HandleFunc("/projects", l.Log(shared.Validate(p.Create(t.End()))).ServeHTTP).Methods("POST")
+	r.HandleFunc("/projects/{id}", l.Log(shared.Validate(p.Delete(t.End()))).ServeHTTP).Methods("DELETE")
+	r.HandleFunc("/projects/{id}/cancel", l.Log(shared.Validate(p.Cancel(t.End()))).ServeHTTP).Methods("POST")
 
 	http.Handle("/", r)
 	return &HTTPAdapter{
